start github api process
